Add words from official wordle answers.
Previously, wordle-words used words from /usr/share/dict/words matching
`[a-z]{5}`. But some words, like sauté, don't match that pattern but are
in the official answers list without the accent (saute).

This combines /usr/share/dict/words with the official answers list to
make sure anything that can show up as the answer is counted as a
possible answer.